+++ NEWS.md
@@ -1,5 +1,6 @@
-## News for Package Rdimtools
-## Changes in version 0.4.3
+# News for Package Rdimtools
+### Changes in version 0.4.3
+  * LDA solves trace ratio problem directly.
   * Supports
     - 125 dimension reduction methods
     - 13  intrinsic dimension estimation techniques